fix(jit/hir): wrap INC/DEC fold; softcode max/min/sign/bound float (#1259, #1260)
#1259: HIR_INC/HIR_DEC const-fold used signed +1/-1 (C++ UB at
INT64_MAX/MIN). Fold and try_fold now use unsigned intermediates so
wrap matches ADDI and overflow_inject TC005.

#1260: Softcode sign/max/min/bound are float in the interpreter
(mux_atof/fval) but still emitted integer HIR when is_int/all_int.
Stop those native lowers (same class as #1150 abs); constants keep
try_fold float, runtime falls through to ECALL. Residual HIR_SIGN/
MAX/MIN codegen remains for any non-softcode residual.

+++ mux/modules/engine/hir_opt.cpp
@@ -334,10 +334,14 @@ void hir_const_fold(hir_program &h) {
                 break;
 
             // INC / DEC of ICONST.
+            // Signed overflow is C++ UB; softcode/runtime wrap via
+            // ADDI (two's complement).  Fold with unsigned intermediate
+            // so INT64_MAX+1 / INT64_MIN-1 match wrap (#1259).
             case HIR_INC:
                 if (s1 >= 0 && h.kind[s1] == HIR_ICONST) {
                     h.kind[i] = HIR_ICONST;
-                    h.val[i] = h.val[s1] + 1;
+                    h.val[i] = static_cast<int64_t>(
+                        static_cast<uint64_t>(h.val[s1]) + 1u);
                     h.src1[i] = -1;
                     changed = true;
                 }
@@ -345,7 +349,8 @@ void hir_const_fold(hir_program &h) {
             case HIR_DEC:
                 if (s1 >= 0 && h.kind[s1] == HIR_ICONST) {
                     h.kind[i] = HIR_ICONST;
-                    h.val[i] = h.val[s1] - 1;
+                    h.val[i] = static_cast<int64_t>(
+                        static_cast<uint64_t>(h.val[s1]) - 1u);
                     h.src1[i] = -1;
                     changed = true;
                 }
